version 3 & Ready to build CICD

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,5 +1,8 @@
 package com.example.demo.student;
 
+import com.example.demo.exceptions.ApiException;
+import com.example.demo.exceptions.EmailExistException;
+import com.example.demo.exceptions.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,10 +19,18 @@ public class StudentService {
 
     public void addStudent(Student student){
         // check email is taken
+        Boolean emailExists = studentRepository.selectExistsByEmail(student.getEmail());
+        if (emailExists){
+            throw new EmailExistException("Email " + student.getEmail() + " already exists!");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId){
+        // check studentId
+        if (!studentRepository.existsById(studentId)){
+            throw new StudentNotFoundException("Student with student id " + studentId + " does not exist!");
+        }
         studentRepository.deleteById(studentId);
     }
 }
